# docs/api-reference/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/{network}/submit": {
      "post": {
        "operationId": "CloseController_submit",
        "parameters": [
          {
            "name": "network",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "testnet",
                "mainnet"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Confirmed: returns the transaction hash and ledger.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid network, or invalid or unsigned/undecodable transaction envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "invalid_network",
                                "invalid_signed_xdr",
                                "invalid_signature",
                                "invalid_body"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "unauthorized"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "413": {
            "description": "The request body is larger than 100 KB.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "payload_too_large"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "The request body is not UTF-8 JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "unsupported_media_type"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for this key.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "rate_limited"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "502": {
            "description": "The network rejected the transaction.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "submit_rejected",
                                "submit_failed"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "504": {
            "description": "The transaction did not confirm in time.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "confirmation_timeout"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Submit a client-signed transaction and wait for confirmation.",
        "tags": [
          "close"
        ]
      }
    },
    "/v1/{network}/stats": {
      "get": {
        "operationId": "StatsController_totals",
        "parameters": [
          {
            "name": "network",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "testnet",
                "mainnet"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsTotalsDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid network.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "invalid_network"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "The stats store is unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "stats_unavailable"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Accounts closed and XLM recovered (public, no API key).",
        "tags": [
          "stats"
        ]
      }
    },
    "/v1/{network}/stats/feed": {
      "get": {
        "operationId": "StatsController_feed",
        "parameters": [
          {
            "name": "network",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "testnet",
                "mainnet"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsFeedDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid network.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "invalid_network"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "The stats store is unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "stats_unavailable"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Totals, recent closes and daily activity (public, no API key).",
        "tags": [
          "stats"
        ]
      }
    },
    "/v1/{network}/stats/merges": {
      "post": {
        "description": "Closes submitted through `POST /v1/{network}/submit` are counted automatically. The transaction is re-read from the network and counted only if it confirmed and merged an account; repeating a hash never counts it twice.",
        "operationId": "StatsController_record",
        "parameters": [
          {
            "name": "network",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "testnet",
                "mainnet"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordMergeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordMergeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid network, transaction hash or JSON body, or the merge was not verified on the network.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "invalid_network",
                                "invalid_tx_hash",
                                "tx_not_verified",
                                "invalid_body"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "unauthorized"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "413": {
            "description": "The request body is larger than 100 KB.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "payload_too_large"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "The request body is not UTF-8 JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "unsupported_media_type"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for this key.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "rate_limited"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "The stats store is unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "stats_unavailable"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Count a confirmed close submitted outside this API.",
        "tags": [
          "stats"
        ]
      }
    },
...
  },
  "info": {
    "title": "LumenWipe API",
    "description": "Programmatic close-out of Stellar accounts. Request bodies must be JSON objects of at most 100 KB: a larger body is rejected with 413 payload_too_large, a non-JSON content type with 415 unsupported_media_type, and a body that is not a JSON object with 400 invalid_body.",
    "version": "0.1.0",
    "contact": {
      "name": "LumenWipe",
      "url": "https://lumenwipe.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "servers": [
    {
      "url": "https://api.lumenwipe.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "Integrator API key"
      },
      "integrator-session": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "Short-lived wallet session token from POST /integrator/auth/session"
      }
    },
    "schemas": {
...
      "ErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "account_not_found",
              "account_read_failed",
              "account_too_large",
              "admin_api_not_configured",
              "allowances_read_failed",
              "api_key_not_found",
              "aquarius_trustline_missing",
              "backstop_emissions_unclaimed",
              "backstop_token_unconvertible",
              "backstop_withdrawal_cooling_down",
              "backstop_withdrawal_not_queued",
              "bad_request",
              "blend_emissions_trustline_missing",
              "blend_repay_asset_balance_unknown",
              "blend_repay_asset_missing",
              "confirmation_timeout",
              "conversion_floor_missing",
              "conversion_provider_unrecognized",
              "defi_exit_blocked",
              "defi_exit_unsupported",
              "defi_position_unrecognized",
              "defi_positions_blocked",
              "defi_positions_stale",
              "defi_positions_unavailable",
              "destination_not_acknowledged",
              "fee_bump_exceeds_cap",
              "fee_bump_not_configured",
              "forbidden",
              "forward_amount_exceeds_balance",
              "inner_fee_not_zero",
              "integrator_api_not_configured",
              "internal_error",
              "invalid_address",
              "invalid_addresses",
              "invalid_body",
              "invalid_challenge",
              "invalid_decisions",
              "invalid_destination",
              "invalid_memo",
              "invalid_network",
              "invalid_owner",
              "invalid_rate_limit",
              "invalid_request",
              "invalid_signature",
              "invalid_signed_xdr",
              "invalid_source",
              "invalid_spender",
              "invalid_token",
              "invalid_tokens",
              "invalid_transaction_xdr",
              "invalid_tx_hash",
              "key_limit_reached",
              "mediator_key_misconfiguration",
              "mediator_not_configured",
              "memo_required",
              "merged_account_not_found",
              "missing_parameters",
              "missing_transaction",
              "needs_decisions",
              "not_found",
              "operation_not_sponsorable",
              "owner_not_found",
              "path_lookup_failed",
              "payload_too_large",
              "phoenix_trustline_missing",
              "plan_failed",
              "provider_response_unusable",
              "quote_drifted",
              "rate_limited",
              "registry_expired",
              "request_failed",
              "revoke_build_failed",
              "revoke_needs_restore",
              "revoke_simulation_failed",
              "revoke_unsafe",
              "service_unavailable",
              "signer_normalization_unsafe",
              "soroban_token_conversion_failed",
              "soroban_token_conversion_unavailable",
              "soroban_token_conversion_unsafe",
              "soroban_token_needs_restore",
              "soroban_token_route_lost",
              "soroban_token_transfer_failed",
              "soroban_token_transfer_unsafe",
              "soroban_token_unreadable",
              "soroswap_trustline_missing",
              "stats_unavailable",
              "submit_failed",
              "submit_rejected",
              "too_many_addresses",
              "transaction_structure_not_allowed",
              "transactions_failed",
              "transfer_destination_missing",
              "transfer_destination_unusable",
              "trustline_cannot_be_left",
              "trustline_deauthorized_with_balance",
              "tx_not_verified",
              "unauthorized",
              "unprocessable_entity",
              "unsupported_media_type",
              "unsupported_memo_type",
              "vault_undercollateralized",
              "withdraw_before_repay"
            ],
            "description": "Stable machine-readable code. Branch on this, never on the message."
          },
          "message": {
            "type": "string",
            "description": "Plain-language explanation. May be reworded."
          },
          "details": {
            "type": "object",
            "description": "Extra structured context, present only when there is more to say."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SubmitRequestDto": {
        "type": "object",
        "properties": {
          "signedXdr": {
            "type": "string",
            "description": "Signed transaction envelope (base64 XDR) produced by the client.",
            "example": "AAAAAgAAAAB..."
          }
        },
        "required": [
          "signedXdr"
        ]
      },
      "SubmitResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success"
            ],
            "example": "success"
          },
          "hash": {
            "type": "string",
            "description": "The submitted transaction's hash.",
            "example": "a1b2c3...d4e5f6"
          },
          "ledger": {
            "type": "number",
            "description": "The ledger the transaction was included in.",
            "example": 123456
          }
        },
        "required": [
          "status",
          "hash",
          "ledger"
        ]
      },
      "StatsTotalsDto": {
        "type": "object",
        "properties": {
          "network": {
            "type": "string",
            "enum": [
              "testnet",
              "mainnet"
            ],
            "example": "mainnet"
          },
          "accountsClosed": {
            "type": "number",
            "description": "Accounts closed through LumenWipe.",
            "example": 1284
          },
          "xlmRecoveredStroops": {
            "type": "string",
            "description": "XLM recovered by those closes, in stroops, as a decimal string.",
            "example": "52310000000"
          }
        },
        "required": [
          "network",
          "accountsClosed",
          "xlmRecoveredStroops"
        ]
      },
      "MergeRecordDto": {
        "type": "object",
        "properties": {
          "txHash": {
            "type": "string",
            "example": "3389e9f0f1a65f19736cacf544c2e825313e8447f569233bb8db39aa607c8889"
          },
          "network": {
            "type": "string",
            "enum": [
              "testnet",
              "mainnet"
            ],
            "example": "mainnet"
          },
          "accountsClosed": {
            "type": "number",
            "description": "Accounts this transaction merged.",
            "example": 1
          },
          "xlmStroops": {
            "type": "string",
            "description": "XLM the merges moved, in stroops.",
            "example": "45000000"
          },
          "timestamp": {
            "type": "string",
            "description": "Ledger close time, ISO 8601.",
            "example": "2026-10-08T14:03:11.000Z"
          }
        },
        "required": [
          "txHash",
          "network",
          "accountsClosed",
          "xlmStroops",
          "timestamp"
        ]
      },
...
      "StatsFeedDto": {
        "type": "object",
        "properties": {
          "totals": {
            "$ref": "#/components/schemas/StatsTotalsDto"
          },
          "recent": {
            "description": "The 50 most recent closes, newest first.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MergeRecordDto"
            }
          },
          "daily": {
            "description": "The last 365 UTC days, oldest first, zero-filled.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyActivityDto"
            }
          }
        },
        "required": [
          "totals",
          "recent",
          "daily"
        ]
      },
      "RecordMergeRequestDto": {
        "type": "object",
        "properties": {
          "txHash": {
            "type": "string",
            "description": "Hash of a confirmed transaction that merged the account being closed.",
            "example": "3389e9f0f1a65f19736cacf544c2e825313e8447f569233bb8db39aa607c8889"
          }
        },
        "required": [
          "txHash"
        ]
      },
      "RecordMergeResponseDto": {
        "type": "object",
        "properties": {
          "counted": {
            "type": "boolean",
            "description": "True when this call counted the close; false when the hash was already counted.",
            "example": true
          }
        },
        "required": [
          "counted"
        ]
      },
...
    }
  }
}